API to generate UUID
